+++ repomancer.go
@@ -4,30 +4,25 @@ import (
 	"fyne.io/fyne/v2/app"
 	"fyne.io/fyne/v2/cmd/fyne_demo/data"
 	"github.com/repomancer/repomancer/internal"
-	"github.com/repomancer/repomancer/screen"
 	"log"
 	"os"
+
+	"github.com/repomancer/repomancer/ui"
 )
 
 func main() {
 	a := app.NewWithID("com.sheersky.repomancer")
 	a.SetIcon(data.FyneLogo)
-	screens.LogLifecycle(a)
-	w := a.NewWindow("Repomancer")
-
-	w.SetMainMenu(screens.MakeMenu(a, w))
-	w.SetMaster()
-
+	myUi := ui.NewBaseUI(a)
 	if len(os.Args) < 2 {
-		screens.GotoStartScreen(a, w)
+		myUi.ShowWindow(ui.Start)
 	} else {
 		project, err := internal.OpenProject(os.Args[1])
 		if err != nil {
 			log.Fatal(err)
 		}
-		screens.GotoProjectScreen(w, project)
-
+		myUi.ShowProjectWindow(project)
 	}
-	w.ShowAndRun()
+	a.Run()
 
 }