+++ ui/dialogs/pullrequestdialog.go
@@ -1,4 +1,4 @@
-package screens
+package dialogs
 
 import (
 	"fmt"
@@ -6,7 +6,7 @@ import (
 	"fyne.io/fyne/v2/dialog"
 	"fyne.io/fyne/v2/widget"
 	"github.com/repomancer/repomancer/internal"
-	"github.com/repomancer/repomancer/screen/widgets"
+	"github.com/repomancer/repomancer/ui/widgets"
 	"strings"
 )